1. Modified README.md about Pedal
2. Modified Libraries in pedal_pico
 a. Cleaned Up
 b. Removed Output Modes of pedal_pico_sideband
 c. Removed Loss of pedal_pico_distortion

3. Modified util_pedal_pico
 a. Fixed PWM Output (Balanced Monaural)
 b. Fixed Function Not to Get Warning in Compiling

+++ include/util_pedal_pico.h
@@ -168,7 +168,7 @@ void util_pedal_pico_wait();
 void util_pedal_pico_init_multi(uchar8 gpio_bit_0, uchar8 gpio_bit_1, uchar8 gpio_bit_2, uchar8 gpio_bit_3);
 void util_pedal_pico_select_multi();
 void (**util_pedal_pico_multi_set)();
-void (**util_pedal_pico_multi_process)(uint16, uint16, uint16, uchar8);
+void (**util_pedal_pico_multi_process)(int32, uint16, uint16, uchar8);
 void (**util_pedal_pico_multi_free)();
 /**
  * Caution! These Functions tries to erase and write data to the on-board flash memory.
